GameJam-submitted version!
The commit includes the asset pack assets that made it to the sumbitted version of the game. Also includes music and a sound effect for the rolling "death ball" trap.
Added a Tutorial and Credits menu as well.
Fixed the occasional bug here and there.
This version does not include pre-computed lighting to help with file size (frustratingly it didn't actually reduce the size nearly at all!), and it does have a bit of an impact in the interior scene which looks a bit brighter and less atmospheric as the intended builds earlier.
Future versions may arrive that include the original intended lighting/shading for the interior part of the map.