Renamed the blue tile material for the platform (its suffix was wrong!) Finished the implementation of the TriggerPlatforms in C++. The Blueprint actor now only handles the timeline for the movement of the platform's mesh. Created an EditorOnly functionality that allows me to pick the PuzzleElementType of the TriggerPlatforms and the material on the mesh changes automatically to blue/red/green. Had to change the component hieararchy in the TriggerPlatform. In the Brazier asset, made the light to be a child of the FireComponent.
ISSUE: The resulting changes in the TriggerPlatform error now causes some issues with the moving platform assets that spawn a triggerplatform on them, but that's to be expected.
Made a minor correction with the platform mesh reference, but they will be fixed more properly when refactoring of the moving platform blueprits is complete later on.